Data.Ranges

Documentation

range :: Ord a => a -> a -> Range a

Construct a Range from a lower and upper bound.

ranges :: Ord a => [Range a] -> Ranges a

Construct a Ranges from a list of lower and upper bounds.

data Ord a => Range a

data Ord a => Ranges a

inRange :: Ord a => a -> Range a -> Bool

Tests if a given range contains a particular value.

inRanges :: Ord a => a -> Ranges a -> Bool

Tests if any of the ranges contains a particular value.

toSet :: Ord a => Ranges a -> Set (Range a)

Allows quick lookups using ranges.

single :: Ord a => a -> Range a

A range consisting of a single value.

addRange :: Ord a => Ranges a -> Range a -> Ranges a